fix(root package): Renamed the package name to com.ntnu.idatt
Reason:
Because the package name should not end in numbers. The previous package
name was com.ntnu.idatt2003. It led to mvn clean test getting errors and
CI failed
