+++ src/main/java/com/ntnu/idatt/entity/Player.java
@@ -1,4 +1,4 @@
-package com.ntnu.idatt2003.entity;
+package com.ntnu.idatt.entity;
 
 /*
 import org.slf4j.Logger;
@@ -6,8 +6,8 @@ import org.slf4j.LoggerFactory;
  */
 
 
-import com.ntnu.idatt2003.logic.BoardGame;
-import com.ntnu.idatt2003.model.Tile;
+import com.ntnu.idatt.logic.BoardGame;
+import com.ntnu.idatt.model.Tile;
 
 /**
  * The type Player.
