+++ src/main/java/com/ntnu/idatt/logic/action/TileAction.java
@@ -1,0 +1,8 @@
+package com.ntnu.idatt.logic.action;
+
+import com.ntnu.idatt.entity.Player;
+
+public interface TileAction {
+  
+  void perform(Player player);
+}
