--- src/main/java/com/ntnu/idatt2003/logic/action/TileAction.java
@@ -1,8 +1,0 @@
-package com.ntnu.idatt2003.logic.action;
-
-import com.ntnu.idatt2003.entity.Player;
-
-public interface TileAction {
-  
-  void perform(Player player);
-}
