+++ src/test/java/com/ntnu/idatt/model/BoardTest.java
@@ -1,6 +1,4 @@
-package com.ntnu.idatt2003.model;
-
-import static org.junit.jupiter.api.Assertions.*;
+package com.ntnu.idatt.model;
 
 import org.junit.jupiter.api.Test;
